==> Code_Notebooks/1_Data_Collection.ipynb <==
# Data Collection

# imports
import pandas as pd
import numpy as np

## MovieLens Data Structure
This data from the [MovieLens 25M Dataset](https://grouplens.org/datasets/movielens/25m/). MovieLens itself is an online recommender system that allows its users to add ratings and tags to movies. This data includes reviews and tags taken directly from MovieLens users, as well as movie data organized by unique movie IDs. This dataset from MovieLens was collected in 2019 from 162,00 users. It contains 25 million ratings and one million tags for 62,000 movies.

### Movies Data Set
This data set pairs each MovieLens ID with its title, formatted 'Name of Movie (Year)', and its genres, formatted 'genre1|genre2|...'.

movies = pd.read_csv('../Data/Large-Data/MovieLens-25M/movies.csv')
movies.head()

movies.tail()

movies.isna().sum()

# This dataset covers 62,423 movies
movies.info()

# Lets check there are no duplicates
print(f"There are {movies['movieId'].nunique()} unique movie IDs.")
print(f"There are {movies['title'].nunique()} unique titles.")

There are 62,423 unique movie IDs, but only 62,325 unique titles. Since `title` has the name and year this may mean we have movies with the same title and year or it could mean duplicates.

# It looks like the most often a movie pops up is twice. Since there are unique movie IDs, we will leave this for now, but keep in mind during EDA and modeling
# Since the most a title occurs is twice, that means we only have 98 duplicates out of over 62,000
movies['title'].value_counts()

We'll need to figure out how to better organize `genres`, most likely with boolean columns with each genre, for example `genres_fantasy`. It would be interesting to look at how movies that share multiple genres interact or if certain people like movies that have multiple genres more than movies that only have one. 

### Links Data Set
This data set pairs each MovieLens ID with its IMDb and TMDB IDs. This is needed for combining our data from IMDb with our MovieLens data.

links = pd.read_csv('../Data/Large-Data/MovieLens-25M/links.csv')
links.head()

print(f'The shape of the links DataFrame is {links.shape}')
print(f"There are {links['movieId'].nunique()} unique MovieLens IDs.")
print(f"There are {links['imdbId'].nunique()} unique IMDb IDs.")
print(f"There are {links['tmdbId'].nunique()} unique TMDB IDs.")

# Let's check for any missing values
links.isna().sum()

We have the same number of unique MovieLens IDs as our `movies` DataFrame, which also matches our unique IMDb IDs. This should make it easy to match the IMDb and MovieLens IDs up when combining our reviews from each into one dataframe. There are some duplicate TMDB IDs as well as some missing values, but as of right now we will not need to use that column, so we can ignore this.

### Ratings Data Set
This data contains the ratings from MovieLens users, up to 2019. It has mulitple reviews per userId, broken up by movieId, rating, and timestamp.

ratings = pd.read_csv('../Data/Large-Data/MovieLens-25M/ratings.csv')
ratings.head()

print(f'The shape of the ratings DataFrame is {ratings.shape}')
print(f"There are {ratings['userId'].nunique()} unique MovieLens user IDs.")
print(f"There are {ratings['movieId'].nunique()} unique movies rated.")

Only 59,047 movies out of the 62,000 that we have titles for have ratings. We will need to keep this in mind when considering which movies to include in our recommender system, as we don't want to be recommending movies without having any information on them.

ratings.dtypes

# Let's check for missing values
ratings.isna().sum()

# There are 10 possible ratings on a scale of 0.5-5
print(ratings['rating'].nunique())
print(ratings['rating'].unique())

Our data for these ratings looks pretty clean. We will need to scale the ratings to match the IMDb ratings we have, but there are no incorrect data types or missing values.

### Tags Data Set
The [MovieLens Data Dictionary](https://files.grouplens.org/datasets/movielens/ml-25m-README.html) defines tags as "user-generated metadata about movies. Each tag is typically a single word or short phrase." These tags are user generated and added to movies on the MovieLens database. This data set is a set of unique user tags, containing the MovieLens user ID, MovieLens movie ID, tag, and timestamp of the tag.

tags = pd.read_csv('../Data/Large-Data/MovieLens-25M/tags.csv')
tags.head()

# We have over 1,000,000 tags, about 73,000 of them unique
print(f"The shape of the tags DataFrame is {tags.shape}.")
print(f"There are {tags['tag'].nunique()} unique tags.")
print(f"There are tags for {tags['movieId'].nunique()} unique movies.")

Only 45,251 of our 62,000 movies have tags. We may want to consider dropping movies that do not have tags or ratings from our data, but we will examine that when we combine the data.

# Let's look for any missing values
tags.isna().sum()

Since we are only interesting in the tags column, a row with no tag is of no use to us. We can drop these rows without worrying, since it is only decreasing our data by 16 rows out of more than 1,000,000.

## IEEE IMDb Review Data
This data set comes from [IEEE DataPort](https://ieee-dataport.org/open-access/imdb-users-ratings-dataset). It contains 4,669,820 ratings from 1,499,238 users to 351,109 movies on IMDb. Each row has a userID, movieID, rating, and review date. Instructions were included

# This code comes directly from the IEEE instructions for reading in this data

dataset = np.load ("../Data/Large-Data/imdb_reviews.npy")

dataset[0]

It is set up as a numpy array so we need to reformat it to turn it into a DataFrame

# Turn into list or rows seperated by commas, like a csv file
data_list = [row.split(',') for row in dataset]

# Convert to pandas DataFrame with labeled columns
imdb_ratings = pd.DataFrame(data_list, columns=['UserID', 'MovieID', 'Rating', 'ReviewDate'])

print(imdb_ratings.dtypes)
imdb_ratings.head()

# Convert Rating to numeric
imdb_ratings['Rating'] = pd.to_numeric(imdb_ratings['Rating'])

# Convert ReviewDate to datetime
imdb_ratings['ReviewDate'] = pd.to_datetime(imdb_ratings['ReviewDate'])

imdb_ratings.dtypes

print(f'The shape of the IMDb ratings DataFrame is {imdb_ratings.shape}')
print(f"There are {imdb_ratings['UserID'].nunique()} unique IMDb user IDs.")
print(f"There are {imdb_ratings['MovieID'].nunique()} unique movies rated.")

There are fewer ratings than the MovieLens data, but they cover almost 300,000 more movies. This will be helpful to make sure we have enough data and may prevent us from having to remove any movies from our recommender system.

# These ratings are on a scale of 1-10, unlike the MovieLens which is 0.5 to 5
# We can scale these later to match, since there are 10 options for each
print(imdb_ratings['Rating'].nunique())
print(imdb_ratings['Rating'].unique())

# Lets save this as a csv so we don't need to reformat it again
imdb_ratings.to_csv('../Data/Large-Data/ieee_imdb_reviews.csv', index=False)

imdb_reviews = pd.read_csv('../Data/Large-Data/ieee_imdb_reviews.csv')
imdb_reviews.head()

## Kaggle Movies Data

### Movie Metadata

movies_metadata = pd.read_csv('../Data/Large-Data/kaggle-movies/movies_metadata.csv')
movies_metadata

movies_metadata.isnull().sum()

credits = pd.read_csv('../Data/Large-Data/kaggle-movies/credits.csv')
credits

keywords = pd.read_csv('../Data/Large-Data/kaggle-movies/keywords.csv')
keywords

links_small = pd.read_csv('../Data/Large-Data/kaggle-movies/links_small.csv')
links_small

ratings_small = pd.read_csv('../Data/Large-Data/kaggle-movies/ratings_small.csv')
ratings_small

credits = pd.read_csv('../Data/Large-Data/kaggle-movies/credits.csv')
credits

## IMDb Title Data

imdb_titles = pd.read_csv('../Data/Large-Data/imdb_title_akas.tsv', sep='\t')
imdb_titles

imdb_titles.dtypes

originals = imdb_titles[imdb_titles['types']=='original']
originals

originals.isna().sum()

originals = originals.dropna()

print(originals.shape)
print(originals['titleId'].nunique())

originals = originals.drop_duplicates(subset=['titleId'])
print(originals.shape)
print(originals['titleId'].nunique())

originals[originals['titleId']=='tt0114709']

imdb_titles = originals[['titleId', 'title']]
imdb_titles

imdb_titles.to_csv('../Data/Large-Data/imdb_titles.csv', index=False)
