==> Code_Notebooks/2_Data_Cleaning.ipynb <==
# Data Cleaning
In this notebook, we will be cleaning the data sets that were reviewed in the Data Collection notebook. For modeling we need to combine the features we want to use into on dataframe for modeling.
<br><br> This includes:
 - Combining MovieLens and IMDb reviews into one dataframe
 - Filtering on the number of reviews by user

# Imports
import pandas as pd

## Combining MovieLens and IMDb Reviews
### Reformatting Movie IDs
Our data from MovieLens comes with three different IDs for each movie, but we will be merging our data using the IMDb ID from our IMDb review data. The MovieLens IDs only have a number for IMDb ID, whereas the other data sets have it formatted as an 9-character code starting with "tt". Toy Story, for example, is "tt0114709" in the IMDB data and only "114709" in the MovieLens Data.

ml_links = pd.read_csv('../Data/Large-Data/MovieLens-25M/links.csv')
ml_links.head()

ml_links.sort_values(by = 'imdbId', ascending=True)

# Lets add as a new column called 'imdb_reformatted'
ml_links['imdb_reformatted'] = 'tt' + ml_links['imdbId'].astype(str).str.zfill(7)

ml_links.sort_values(by = 'imdbId', ascending=True)

### Reformatting MovieLens Ratings
Our MovieLens ratings are scored out of 5, with 10 possible scores. The IMDb scores also have 10 options, but are formatted as 1-10. We can scale these to match so we can combine them into one set of ratings.

ml_ratings = pd.read_csv('../Data/Large-Data/MovieLens-25M/ratings.csv')
ml_ratings.head()

# Multiply the MovieLens rating by 2 so we have the same 1-10 rating scale as IMDb
ml_ratings['scaled_rating'] = ml_ratings['rating'] * 2
ml_ratings.head()

# We need to add our new IMDb ID column to our MovieLens ratings
ml_ratings_for_merge = ml_ratings[['userId', 'movieId', 'scaled_rating']]
ml_ratings_for_merge

ml_ratings_for_merge = pd.merge(ml_ratings_for_merge, ml_links, on='movieId', how='left')
ml_ratings_for_merge

# We need to rename our columns to match the MovieLens reviews
ml_ratings_for_merge = ml_ratings_for_merge.rename({'userId': 'user_id', 'imdb_reformatted': 'imdb_id'}, axis='columns')
ml_ratings_for_merge.head()

# Take only the columns we will use for modeling
ml_ratings_for_merge = ml_ratings_for_merge[['user_id', 'imdb_id', 'scaled_rating']]
ml_ratings_for_merge

### Reformatting IMDb Reviews
We need to rename our IMDb review columns to match before we can merge with the MovieLens reviews.

# read in our IMDb reviews
imdb_reviews = pd.read_csv('../Data/Large-Data/ieee_imdb_reviews.csv')
imdb_reviews.head()

# Drop unneeded columns
imdb_reviews.drop(columns='ReviewDate', inplace=True)

# We need to rename our columns to match the MovieLens reviews
imdb_reviews = imdb_reviews.rename({'UserID': 'user_id', 'MovieID': 'imdb_id', 'Rating': 'scaled_rating'}, axis='columns')
imdb_reviews.head()

### Combining into one DataFrame
We will combine these reviews into one dataframe to have a unified set of reviews. We need to rename our columns to match and then we can concatonate vertically.

merged_reviews = pd.concat([ml_ratings_for_merge, imdb_reviews], axis=0)
merged_reviews

# Let's double check that the shape of our merged reviews is correct
print(ml_ratings_for_merge.shape)
print(imdb_reviews.shape)
print(merged_reviews.shape)

### Adding in Title from IMDb Title Data
We will need the titles to show in our recommender system. We will merge them over from our IMDb Title data set

imdb_titles = pd.read_csv('../Data/Large-Data/imdb_titles.csv')
imdb_titles.head()

imdb_titles = imdb_titles.rename({'titleId': 'imdb_id'}, axis='columns')
imdb_titles.head()

merged_reviews = pd.merge(merged_reviews, imdb_titles, on= 'imdb_id', how='left')
merged_reviews

merged_reviews.isna().sum()

rows_with_null_title = merged_reviews[merged_reviews['title'].isnull()]

rows_with_null_title

Just from quickly looking up these IMDb IDs on IMDb, some of these appear to be TV shows, other are not listed. I think it is safe to remove these since they only account for about 1% of our reviews.

# Let's drop the columns without titles
merged_reviews = merged_reviews.dropna()
merged_reviews.isna().sum()

# Let's save this to use for our recommender system
merged_reviews.to_csv('../Data/Large-Data/merged_reviews.csv', index = False)

## DataFrame Filtered by number of reviews

merged_reviews_filter = pd.read_csv('../Data/Large-Data/merged_reviews.csv')
merged_reviews_filter.head()

print(merged_reviews_filter.shape)

merged_reviews_filter['imdb_id'].nunique()

# Let's filter for movies that have more than 1 review
more_than_1 = merged_reviews_filter.groupby('imdb_id').filter(lambda x: len(x)>1)


# Let's filter for movies that have more than 5 reviews
more_than_5 = merged_reviews_filter.groupby('imdb_id').filter(lambda x: len(x)>5)
len(more_than_5['imdb_id'])

# Let's filter for movies that have more than 10 reviews
more_than_10 = merged_reviews_filter.groupby('imdb_id').filter(lambda x: len(x)>10)


# Let's filter for movies that have more than 20 reviews
more_than_20 = merged_reviews_filter.groupby('imdb_id').filter(lambda x: len(x)>20)


# Let's filter for movies that have more than 100 reviews
more_than_100 = merged_reviews_filter.groupby('imdb_id').filter(lambda x: len(x)>100)

# Let's check how much data we have for movies for each filter
print(f"More than 1: {len(more_than_1['imdb_id'])} reviews for {more_than_1['imdb_id'].nunique()} movies from {more_than_1['user_id'].nunique()} users")
print(f"More than 5: {len(more_than_5['imdb_id'])} reviews for {more_than_5['imdb_id'].nunique()} movies from {more_than_5['user_id'].nunique()} users")
print(f"More than 10: {len(more_than_10['imdb_id'])} reviews for {more_than_10['imdb_id'].nunique()} movies from {more_than_10['user_id'].nunique()} users")
print(f"More than 20: {len(more_than_20['imdb_id'])} reviews for {more_than_20['imdb_id'].nunique()} movies from {more_than_20['user_id'].nunique()} users")
print(f"More than 100: {len(more_than_100['imdb_id'])} reviews for {more_than_100['imdb_id'].nunique()} movies from {more_than_100['user_id'].nunique()} users")

# Let's save this to use for our recommender system
more_than_100.to_csv('../Data/Large-Data/over_100_reviews.csv', index = False)

## Filtering by number of reviews by user

reviews_min_1_per_user = more_than_100.groupby('user_id').filter(lambda x: len(x)>1)

reviews_min_5_per_user = more_than_100.groupby('user_id').filter(lambda x: len(x)>5)

reviews_min_10_per_user = more_than_100.groupby('user_id').filter(lambda x: len(x)>10)

reviews_min_15_per_user = more_than_100.groupby('user_id').filter(lambda x: len(x)>15)

# Let's check how much data we have for movies for each filter
print(f"More than 1: {len(reviews_min_1_per_user['imdb_id'])} reviews for {reviews_min_1_per_user['imdb_id'].nunique()} movies from {reviews_min_1_per_user['user_id'].nunique()} users")
print(f"More than 5: {len(reviews_min_5_per_user['imdb_id'])} reviews for {reviews_min_5_per_user['imdb_id'].nunique()} movies from {reviews_min_5_per_user['user_id'].nunique()} users")
print(f"More than 10: {len(reviews_min_10_per_user['imdb_id'])} reviews for {reviews_min_10_per_user['imdb_id'].nunique()} movies from {reviews_min_10_per_user['user_id'].nunique()} users")

###  Creating dataframe for MovieLens reviews only
We are having issues creating a pivot table with all the reviews

ml_ratings_for_merge.head()

ml_ratings_for_merge = pd.merge(ml_ratings_for_merge, imdb_titles, on= 'imdb_id', how='left')
ml_ratings_for_merge

ml_ratings_for_merge.isna().sum()

# Let's drop the columns without titles
ml_ratings_for_merge = ml_ratings_for_merge.dropna()
ml_ratings_for_merge.isna().sum()

# Let's save this to use for our recommender system
ml_ratings_for_merge.to_csv('../Data/Large-Data/ml_reviews.csv', index = False)
