==> Code_Notebooks/4_Modeling.ipynb <==
# Recommender System Modeling
In this notebook, we will work through multiple iterations of our movie recommender system model. First, we will create a simpler model using just the movie reviews from IMDb. This will give us a baseline model that we can potentially improve on by adding more detailed features.

# Imports
import pandas as pd
import numpy as np
from scipy.sparse import csr_matrix
from sklearn.metrics.pairwise import pairwise_distances

## Selecting Data for Recommender System
This recommender system will use the cosine similarity of user ratings to recommend a group of movies. It will be able to receive a movie title as an input and return movie recommendations that are similar to the movie that is searched.

### Choosing data for model
For the sake of computing power, we will only be using a subset the reviews that we have in our recommender system. This same method can be replicated with larger data sets if more computing power is available.

# Let's read in our MovieLens reviews that were formatted in the Data Collection notebook
ml_reviews = pd.read_csv('../Data/Large-Data/ml_reviews.csv')
ml_reviews

# Let's take about half of our ratings by using only reviews from users with an ID below 81000
sampled_reviews = ml_reviews[ml_reviews['user_id']<81000]
sampled_reviews

# Lets see how much data we have in this sample
print(f"Reviews: {len(sampled_reviews['imdb_id'])}")
print(f"Movies: {sampled_reviews['imdb_id'].nunique()}")
print(f"Users: {sampled_reviews['user_id'].nunique()}")

### Titles in Recommender

imdb_titles = pd.read_csv('../Data/Large-Data/imdb_titles.csv')

imdb_titles.head()

movies_in_sample = sampled_reviews.drop_duplicates(subset = 'imdb_id').sort_values(by='imdb_id', ascending=True)[['imdb_id', 'title']]
movies_in_sample

movies_in_sample['title']

# Create a list of titles ordered by IMDb ID for rows and columns of recommender
ordered_titles = sampled_reviews.drop_duplicates(subset = 'imdb_id').sort_values(by='imdb_id', ascending=True)['title']

ordered_titles

## Creating Recommender System DataFrame
The code below for the function `create_matrix(df)` is based on code from [Recommendation System in Python](https://www.geeksforgeeks.org/recommendation-system-in-python/) from GeeksForGeeks.org. The purpose of this function is to skip the step of creating a pivot table to find cosine similarites. The dataset we are working with is too large to create a pivot table with the computing power available. The code below has had variable names changed to match standard sklearn notation and formatting. The variable `movies_in_sample` has been added so we can refer to it when creating the recommender functions later on. Comments have been added to the code for clarity of the functions purpose in this application.


def create_matrix(df):
    # Get the number of unique User IDs and unique IMDb IDs that appear in our data
    n = len(df['user_id'].unique())
    m = len(df['imdb_id'].unique())
     
    # Map IDs to indices
    user_mapper = dict(zip(np.unique(df["user_id"]), list(range(n))))
    movie_mapper = dict(zip(np.unique(df["imdb_id"]), list(range(m))))
     
    # Map indices to IDs
    user_inv_mapper = dict(zip(list(range(n)), np.unique(df["user_id"])))
    movie_inv_mapper = dict(zip(list(range(m)), np.unique(df["imdb_id"])))

    # Get indices for the creation of our sparse matrix
    user_index = [user_mapper[i] for i in df['user_id']]
    movie_index = [movie_mapper[i] for i in df['imdb_id']]
    
    # Create Matrix
    matrix = csr_matrix((df["scaled_rating"], (movie_index, user_index)), shape=(m, n))

    # Create list of unique movies in sample that will be used for the index of the recommender system
    movies_in_sample = sampled_reviews.drop_duplicates(subset = 'imdb_id').sort_values(by='imdb_id', ascending=True)[['imdb_id', 'title']]
     
    return matrix, user_mapper, movie_mapper, user_inv_mapper, movie_inv_mapper, user_index, movie_index, movies_in_sample
     


matrix, user_mapper, movie_mapper, user_inv_mapper, movie_inv_mapper, user_index, movie_index, movies_in_sample = create_matrix(sampled_reviews)

# Confirm that the length match the unique movie IDs
len(movie_mapper)

#movie_mapper

similarities = 1.0 - pairwise_distances(matrix, metric = 'cosine')

similarities.shape

#recommender = pd.DataFrame(similarities, index = ordered_titles, columns=ordered_titles)
recommender = pd.DataFrame(similarities, index = movies_in_sample['title'], columns=movies_in_sample['title'])

recommender.head()

imdb_titles = pd.read_csv('../Data/Large-Data/imdb_titles.csv')

imdb_titles.head()

# Get unique IMDb IDs from our movie 
movie_mapper_ids = set(movie_mapper.keys())

# Assuming `imdb_titles` is your DataFrame with columns 'imdb_id' and 'title'
# Filter the DataFrame based on whether the imdb_id exists in movie_mapper_ids
titles_in_movie_mapper = imdb_titles[imdb_titles['titleId'].isin(movie_mapper_ids)]

# Print the resulting DataFrame
titles_in_movie_mapper.head()

### Getting Recommendations

def get_recommendations():
    x = input('What movie did you just watch?')
    #titles = titles_in_movie_mapper
    titles = movies_in_sample
    for title in titles.loc[titles['title'].str.contains(x), 'title']:
        print(title)
        print(recommender[title].sort_values(ascending = False)[1:11])
        print('='*40)

get_recommendations()

### Getting Recommendations for a Specific Title

def get_recommendations_exact_title():
    x = input('What movie did you just watch?')
    print(f"Movies like: {x}")
    print('='*20)
    print(recommender[x].sort_values(ascending = False)[1:11])

# run function searching for Toy Story
get_recommendations_exact_title()

### Specific Movie Search that can handle Duplicates
The following function is a work in progress. Errors may occur when searching for movies that share a title with another movie if the commented out code is added back in.

len(movies_in_sample[movies_in_sample['title'] == 'dsfsdsdvsd'])

def get_recommendations_exact_title_w_dups():
    #x = input('What movie did you just watch?')
    x = 'Bad Boys'
    if len(movies_in_sample[movies_in_sample['title'] == x]) == 1:
        print(f"Movies like: {x}")
        print('='*20)
        print(recommender[x].sort_values(by = x, ascending = False)[1:11])
    elif len(movies_in_sample[movies_in_sample['title'] == x]) > 1:
        print(f"There are {len(movies_in_sample[movies_in_sample['title'] == x])} movies named {x}.")
        #for title in movies_in_sample.loc[movies_in_sample['title'] == x]:
            #print(title)
            #print(recommender[title].sort_values(ascending = False)[1:11])
    else:
        print(f'Sorry, {x} was not found in our system.')

    

get_recommendations_exact_title_w_dups()
